second commit

+++ app/src/main/java/com/example/inlab/apiresiduo/model/ResiduoResponse.kt
@@ -1,6 +1,7 @@
 package com.example.inlab.apiresiduo.model
+import kotlin.collections.List
 
 data class ResiduoResponse(
-    val cantidadHecho: Long,
+    val cantidadHecho: Int,
     val nombreLogro: List<String>
 )
